fix: offset is uint64

+++ model/model.go
@@ -14,7 +14,7 @@ type InscriptionBRC20Data struct {
 	TxId       string `json:"-"`
 	Idx        uint32 `json:"-"`
 	Vout       uint32 `json:"-"`
-	Offset     uint32 `json:"-"`
+	Offset     uint64 `json:"-"`
 
 	Satoshi  uint64 `json:"-"`
 	PkScript string `json:"-"`
